+++ R/plotTimeseries.R
@@ -40,9 +40,16 @@
 #'
 #' @export
 #'
-plotTimeseries <- function(x, bin='1hour', column, title=NULL, units=NULL,
-                           style=c('line', 'heatmap'), q=0, by=NULL,
-                           cmap=viridis_pal()(25), toTz='UTC') {
+plotTimeseries <- function(x, 
+                           bin='1hour', 
+                           column, 
+                           title=NULL, 
+                           units=NULL,
+                           style=c('line', 'heatmap'),
+                           q=0, 
+                           by=NULL,
+                           cmap=viridis_pal()(25),
+                           toTz='UTC') {
     x <- checkSimple(x, needCols='UTC')
     x$UTC <- with_tz(x$UTC, tzone=toTz)
     x$timeBin <- floor_date(x$UTC, unit=bin)